# packages/api-types/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateMandateRequest": {
        "properties": {
          "client_id": {
            "title": "Client Id",
            "type": "string"
          },
          "due_on": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due On"
          },
          "instructed_on": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instructed On"
          },
          "kind": {
            "enum": [
              "valuation",
              "due_diligence",
              "rera",
              "transaction",
              "portfolio"
            ],
            "title": "Kind",
            "type": "string"
          },
          "purpose": {
            "enum": [
              "loan",
              "ibc",
              "dispute",
              "financial_reporting",
              "internal"
            ],
            "title": "Purpose",
            "type": "string"
          },
          "valuer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valuer Id"
          }
        },
        "required": [
          "client_id",
          "kind",
          "purpose"
        ],
        "title": "CreateMandateRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MandateResponse": {
        "properties": {
          "client_id": {
            "title": "Client Id",
            "type": "string"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "due_on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due On"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "instructed_on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instructed On"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "purpose": {
            "title": "Purpose",
            "type": "string"
          },
          "requires_registered_valuer": {
            "default": false,
            "title": "Requires Registered Valuer",
            "type": "boolean"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "valuer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valuer Id"
          }
        },
        "required": [
          "id",
          "client_id",
          "kind",
          "purpose",
          "status"
        ],
        "title": "MandateResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Real Estate Factory",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/mandates": {
      "get": {
        "operationId": "list_mandates_api_v1_mandates_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MandateResponse"
                  },
                  "title": "Response List Mandates Api V1 Mandates Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Mandates",
        "tags": [
          "mandates"
        ]
      },
      "post": {
        "operationId": "create_mandate_api_v1_mandates_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMandateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MandateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Mandate",
        "tags": [
          "mandates"
        ]
      }
    },
    "/api/v1/mandates/{mandate_id}": {
      "get": {
        "operationId": "get_mandate_api_v1_mandates__mandate_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "mandate_id",
            "required": true,
            "schema": {
              "title": "Mandate Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MandateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Mandate",
        "tags": [
          "mandates"
        ]
      }
    },
...
  }
}